refactor: move all the react components in src/components/
