+++ src/components/coverage-file-table-head.jsx
