+++ src/components/coverage-file-table-row.jsx
