+++ src/components/coverage-meter-bar.jsx
